Move GitContributorsPlugin to Frost.Plugins
