+++ app/Main.hs
@@ -8,26 +8,27 @@ import           Frost.Effects.Python
 import           Frost.Effects.Rholang
 import           Frost.Effects.Stack
 import           Frost.Effects.Sys
-import           Frost.GitContributorsPlugin
-import           Frost.PandocRun               (runInputPandoc, runOutputPandoc)
+import           Frost.PandocRun                     (runInputPandoc,
+                                                      runOutputPandoc)
 import           Frost.Plugin
+import           Frost.Plugins.GitContributorsPlugin
 import           Frost.Plugins.RholangPlugin
 import           Frost.Plugins.StackPlugins
 import           Frost.PythonPlugin
 import           Frost.TimestampPlugin
 import           FrostError
 
-import           Data.Foldable                 (find)
-import           Data.Function                 ((&))
-import qualified Data.Text                     as T
+import           Data.Foldable                       (find)
+import           Data.Function                       ((&))
+import qualified Data.Text                           as T
 import           Polysemy
 import           Polysemy.Error
 import           Polysemy.Trace
 import           PolysemyContrib
-import           System.Environment            (getArgs)
+import           System.Environment                  (getArgs)
 import           System.Exit
 import           System.IO
-import           Text.Pandoc                   (PandocError)
+import           Text.Pandoc                         (PandocError)
 
 main :: IO ()
 main = results >>= traverse handleEithers >>= exit
